Rename lib/csv/util lib/csvutil; adjust paths
